resize int setting

+++ core.py
@@ -109,7 +109,7 @@ class Helper(ctk.CTk):
         setting["app_width"] = app_width
         x, y, width, height = get_screen_size()
 
-        app_width = max(width / 4, app_width)
+        app_width = max(int(width / 4), app_width)
         hwp_width = width - app_width
 
         hwp_x, hwp_y = x, y
